catalog-system/pyproject.toml
[project]
name = "catalog-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.9.1",
    "psycopg2-binary>=2.9.10",
    "watchgod>=0.8.2",
    "werkzeug>=3.1.3",
]
